src/requirements.txt
absl-py==1.0.0
alembic==1.7.6
amqp==5.1.0
astor==0.8.1
atomicwrites==1.4.0
attrs==19.1.0
billiard==3.6.4.0
cachetools==4.2.4
celery== 5.2.0b3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==4.5.4
Flask==2.0.2
Flask-Injector==0.12.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
injector==0.19.0
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.0rc1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
mecab-python3==1.0.3
numpy==1.20
oauthlib==3.2.0
prompt-toolkit==3.0.28
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
style==1.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.8.0
typing==3.7.4.3
typing_extensions==4.1.1
update==0.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
